last tasks

+++ controllers/blogs.js
@@ -48,6 +48,13 @@ blogsRouter.get('/', async (request, response) => {
   })
 
   blogsRouter.delete('/:id', async (request, response) => {
+
+    const decodedToken = jwt.verify(request.tokenExtractor, process.env.SECRET)  
+    if (!request.tokenExtractor || !decodedToken.id) {    
+      return response.status(401).json({ error: 'token missing or invalid' })  
+    }
+    const user = await User.findById(decodedToken.id)
+    console.log(user._id)
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   })
